pillow 8.3 compatibility

+++ nudenet/image_utils.py
@@ -105,7 +105,7 @@ def img_to_array(img, data_format="channels_last", dtype="float32"):
     # Numpy array x has format (height, width, channel)
     # or (channel, height, width)
     # but original PIL image has format (width, height, channel)
-    x = np.asarray(img, dtype=dtype)
+    x = np.asarray(img).astype(dtype)
     if len(x.shape) == 3:
         if data_format == "channels_first":
             x = x.transpose(2, 0, 1)
